requirements.txt
PyYAML==3.11
fpdf==1.7.2
gnupg==2.0.2
psutil==4.1.0
requests==2.9.1
wsgiref==0.1.2
